Protect against rollover

+++ src/lib/St4SerialMaster.h
@@ -77,10 +77,10 @@ class Mst4 : public Stream
       // SHC_CLOCK HIGH for more than 1500 us means that a pair of data bytes is done being exchanged
       #ifdef HAL_SLOW_PROCESSOR
         #define XMIT_TIME 20
-        if ((micros()-lastMicros) < 10000L) return false;
+        if ((long)(micros()-lastMicros) < 10000L) return false;
       #else
         #define XMIT_TIME 40
-        if ((micros()-lastMicros) < 2000L) return false;
+        if ((long)(micros()-lastMicros) < 2000L) return false;
       #endif
 
       uint8_t s_parity=0;
